Conversational-AI-Bot/requirements.txt
streamlit
requests
fastapi
uvicorn[standard]
python-dotenv
python-multipart
SpeechRecognition
gTTS
groq
